+++ src/main/java/org/cryptimeleon/math/structures/groups/debug/DebugGroupImplNoExpMultiExp.java
@@ -9,6 +9,7 @@ import org.cryptimeleon.math.structures.groups.exp.SmallExponentPrecomputation;
 import java.math.BigInteger;
 import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 /**
  * {@link DebugGroupImpl} implementation that counts operations not done inside (multi-)exponentiations
@@ -52,32 +53,10 @@ public class DebugGroupImplNoExpMultiExp extends DebugGroupImpl {
         super(repr);
     }
 
-    /**
-     * Sets the currently used operation count storage bucket to the one with the given name.
-     * If a bucket with the given name does not exist, a new one is created.
-     * <p>
-     * All operations executed after setting a bucket will be counted within that bucket only.
-     * <p>
-     * The name of the default bucket is "default".
-     *
-     * @param name the name of the bucket to enable
-     */
-    @Override
-    public void setBucket(String name) {
-        currentBucket = putBucketIfAbsent(name);
-    }
-
-    /**
-     * Retrieves the bucket from {@code countingBucketMap}, creating a new one if it does not exist yet.
-     * @param name the name of the bucket to retrieve
-     */
-    private CountingBucket putBucketIfAbsent(String name) {
-        return countingBucketMap.computeIfAbsent(name, kName -> new CountingBucket());
-    }
-
     @Override
     public GroupElementImpl exp(GroupElementImpl base, BigInteger exponent, SmallExponentPrecomputation precomputation) {
-        return base.pow(exponent); // this method already counts the exponentiation
+        // this method counts the exponentiation only
+        return base.pow(exponent);
     }
 
     @Override
@@ -88,9 +67,9 @@ public class DebugGroupImplNoExpMultiExp extends DebugGroupImpl {
         for (MultiExpTerm term : mexp.getTerms()) {
             // Use methods where we can disable counting since we only want to count the multi-exponentiation here
             result = (DebugGroupElementImpl) result
-                    .op(((DebugGroupElementImpl) term.getBase()).pow(term.getExponent(),false), false);
+                    .op(((DebugGroupElementImpl) term.getBase()).pow(term.getExponent(), false), false);
         }
-        currentBucket.addMultiExpBaseNumber(mexp.getTerms().size());
+        addMultiExpBaseNumber(mexp.getTerms().size());
         return result;
     }
 
@@ -104,100 +83,28 @@ public class DebugGroupImplNoExpMultiExp extends DebugGroupImpl {
         return true;
     }
 
-
     @Override
-    void incrementNumOps() {
-        currentBucket.incrementNumOps();
-        allBucketsBucket.incrementNumOps();
+    public void setBucket(String name) {
+        currentBucket = putBucketIfAbsent(name);
     }
 
     @Override
-    void incrementNumInversions() {
-        currentBucket.incrementNumInversions();
-        allBucketsBucket.incrementNumInversions();
+    CountingBucket putBucketIfAbsent(String name) {
+        return countingBucketMap.computeIfAbsent(name, kName -> new CountingBucket());
     }
 
     @Override
-    void incrementNumSquarings() {
-        currentBucket.incrementNumSquarings();
-        allBucketsBucket.incrementNumSquarings();
+    CountingBucket getAllBucketsBucket() {
+        return allBucketsBucket;
     }
 
     @Override
-    void incrementNumExps() {
-        currentBucket.incrementNumExps();
-        allBucketsBucket.incrementNumExps();
+    CountingBucket getCurrentBucket() {
+        return currentBucket;
     }
 
     @Override
-    void addMultiExpBaseNumber(int numTerms) {
-        currentBucket.addMultiExpBaseNumber(numTerms);
-        allBucketsBucket.addMultiExpBaseNumber(numTerms);
-    }
-
-    @Override
-    void incrementNumRetrievedRepresentations() {
-        currentBucket.incrementNumRetrievedRepresentations();
-        allBucketsBucket.incrementNumRetrievedRepresentations();
-    }
-
-    @Override
-    long getNumOps(String bucketName) {
-        return putBucketIfAbsent(bucketName).getNumOps();
-    }
-
-    @Override
-    long getNumInversions(String bucketName) {
-        return putBucketIfAbsent(bucketName).getNumInversions();
-    }
-
-    @Override
-    long getNumSquarings(String bucketName) {
-        return putBucketIfAbsent(bucketName).getNumSquarings();
-    }
-
-    @Override
-    long getNumExps(String bucketName) {
-        return putBucketIfAbsent(bucketName).getNumExps();
-    }
-
-    @Override
-    List<Integer> getMultiExpTermNumbers(String bucketName) {
-        return putBucketIfAbsent(bucketName).getMultiExpTermNumbers();
-    }
-
-    @Override
-    long getNumRetrievedRepresentations(String bucketName) {
-        return putBucketIfAbsent(bucketName).getNumRetrievedRepresentations();
-    }
-
-    @Override
-    long getNumOpsAllBuckets() {
-        return allBucketsBucket.getNumOps();
-    }
-
-    @Override
-    long getNumInversionsAllBuckets() {
-        return allBucketsBucket.getNumInversions();
-    }
-
-    @Override
-    long getNumSquaringsAllBuckets() {
-        return allBucketsBucket.getNumSquarings();
-    }
-
-    @Override
-    long getNumExpsAllBuckets() {
-        return allBucketsBucket.getNumExps();
-    }
-
-    @Override
-    List<Integer> getMultiExpTermNumbersAllBuckets() {
-        return allBucketsBucket.getMultiExpTermNumbers();
-    }
-
-    @Override
-    long getNumRetrievedRepresentationsAllBuckets() {
-        return allBucketsBucket.getNumRetrievedRepresentations();
+    Map<String, CountingBucket> getBucketMap() {
+        return countingBucketMap;
     }
 }