+++ src/main/java/org/cryptimeleon/math/structures/groups/debug/DebugGroupImplTotal.java
@@ -5,6 +5,7 @@ import org.cryptimeleon.math.serialization.Representation;
 import java.math.BigInteger;
 import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 /**
  * {@link DebugGroupImpl} implementation that counts operation including those done inside (multi-)exponentiations.
@@ -42,31 +43,6 @@ public class DebugGroupImplTotal extends DebugGroupImpl {
         super(repr);
     }
 
-    /**
-     * Sets the currently used operation count storage bucket to the one with the given name.
-     * If a bucket with the given name does not exist, a new one is created.
-     * <p>
-     * All operations executed after setting a bucket will be counted within that bucket only.
-     * <p>
-     * The name of the default bucket is "default".
-     *
-     * @param name the name of the bucket to enable
-     */
-    @Override
-    public void setBucket(String name) {
-        currentBucket = putBucketIfAbsent(name);
-    }
-
-    /**
-     * Retrieves the bucket with the given name from {@code countingBucketMap},
-     * creating a new one if it does not exist yet.
-     *
-     * @param name the name of the bucket to retrieve
-     */
-    private CountingBucket putBucketIfAbsent(String name) {
-        return countingBucketMap.computeIfAbsent(name, kName -> new CountingBucket());
-    }
-
     @Override
     public boolean implementsOwnExp() {
         return false;
@@ -78,98 +54,27 @@ public class DebugGroupImplTotal extends DebugGroupImpl {
     }
 
     @Override
-    void incrementNumOps() {
-        currentBucket.incrementNumOps();
-        allBucketsBucket.incrementNumOps();
+    public void setBucket(String name) {
+        currentBucket = putBucketIfAbsent(name);
     }
 
     @Override
-    void incrementNumInversions() {
-        currentBucket.incrementNumInversions();
-        allBucketsBucket.incrementNumInversions();
+    CountingBucket putBucketIfAbsent(String name) {
+        return countingBucketMap.computeIfAbsent(name, kName -> new CountingBucket());
     }
 
     @Override
-    void incrementNumSquarings() {
-        currentBucket.incrementNumSquarings();
-        allBucketsBucket.incrementNumSquarings();
+    CountingBucket getAllBucketsBucket() {
+        return allBucketsBucket;
     }
 
     @Override
-    void incrementNumExps() {
-        currentBucket.incrementNumExps();
-        allBucketsBucket.incrementNumExps();
+    CountingBucket getCurrentBucket() {
+        return currentBucket;
     }
 
     @Override
-    void addMultiExpBaseNumber(int numTerms) {
-        currentBucket.addMultiExpBaseNumber(numTerms);
-        allBucketsBucket.addMultiExpBaseNumber(numTerms);
-    }
-
-    @Override
-    void incrementNumRetrievedRepresentations() {
-        currentBucket.incrementNumRetrievedRepresentations();
-        allBucketsBucket.incrementNumRetrievedRepresentations();
-    }
-
-    @Override
-    long getNumOps(String bucketName) {
-        return putBucketIfAbsent(bucketName).getNumOps();
-    }
-
-    @Override
-    long getNumInversions(String bucketName) {
-        return putBucketIfAbsent(bucketName).getNumInversions();
-    }
-
-    @Override
-    long getNumSquarings(String bucketName) {
-        return putBucketIfAbsent(bucketName).getNumSquarings();
-    }
-
-    @Override
-    long getNumExps(String bucketName) {
-        return putBucketIfAbsent(bucketName).getNumExps();
-    }
-
-    @Override
-    List<Integer> getMultiExpTermNumbers(String bucketName) {
-        return putBucketIfAbsent(bucketName).getMultiExpTermNumbers();
-    }
-
-    @Override
-    long getNumRetrievedRepresentations(String bucketName) {
-        return putBucketIfAbsent(bucketName).getNumRetrievedRepresentations();
-    }
-
-    @Override
-    long getNumOpsAllBuckets() {
-        return allBucketsBucket.getNumOps();
-    }
-
-    @Override
-    long getNumInversionsAllBuckets() {
-        return allBucketsBucket.getNumInversions();
-    }
-
-    @Override
-    long getNumSquaringsAllBuckets() {
-        return allBucketsBucket.getNumSquarings();
-    }
-
-    @Override
-    long getNumExpsAllBuckets() {
-        return allBucketsBucket.getNumExps();
-    }
-
-    @Override
-    List<Integer> getMultiExpTermNumbersAllBuckets() {
-        return allBucketsBucket.getMultiExpTermNumbers();
-    }
-
-    @Override
-    long getNumRetrievedRepresentationsAllBuckets() {
-        return allBucketsBucket.getNumRetrievedRepresentations();
+    Map<String, CountingBucket> getBucketMap() {
+        return countingBucketMap;
     }
 }